backend/requirements.txt
Django
django-rest-framework
django-allauth
djangorestframework-simplejwt
django-cors-headers
dj-rest-auth[with_social]
python-decouple
pytz
pyjwt
requests
cryptography